admin finished (probably with bugs)

+++ app/controllers/admin/msu_users_controller.rb
@@ -13,7 +13,7 @@ class Admin::MsuUsersController < ApplicationController
     if @user.save
       redirect_to [:admin, :msu_users]
     else
-      flash.now[:danger] = @user.error.full_messages.to_sentence
+      flash.now[:danger] = @user.errors.full_messages.to_sentence
       render :new
     end
 
@@ -21,16 +21,21 @@ class Admin::MsuUsersController < ApplicationController
 
   def destroy
     user = MsuUser.find(params[:id])
-    suxess_msg = 'Пользователь ' + user.fullname + ' удалён'
-    if user.destroy
-      flash.now[:success] = suxess_msg
+    if user
+      username = user.name
+      if user.destroy
+        flash[:success] = 'Пользователь' + username + ' удалён'
+      else
+        flash[:danger] = @user.error.full_messages.to_sentence
+      end
     else
-      flash.now[:danger] = @user.error.full_messages.to_sentence
+      flash[:danger] = 'Пользователь не найден'
     end
+    redirect_to [:admin, :msu_users]
   end
 
   private
     def user_params
-      params.require(:msu_user).permit(:name, :password)
+      params.require(:msu_user).permit(:name, :password, :admin)
     end
 end
